beta9主要修复的内容：    1.统计    2.客户评价

+++ common/src/main/java/com/easymi/common/entity/WorkStatistics.java
@@ -14,7 +14,7 @@ public class WorkStatistics implements Serializable{
     public int finishCount;
 
     @SerializedName("total_amount")
-    public int totalAmount;
+    public double totalAmount;
 
     public double income;
